requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
Django==4.2
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
environs==9.5.0
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
packaging==23.1
Pillow==9.5.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
